+++ src/main/groovy/ru/vyarus/gradle/plugin/python/task/CheckPythonTask.groovy
@@ -26,6 +26,7 @@ import ru.vyarus.gradle.plugin.python.util.PythonExecutionFailed
  * @since 08.12.2017
  */
 @CompileStatic
+@SuppressWarnings('UnnecessaryGetter')
 class CheckPythonTask extends BasePipTask {
 
     private boolean virtual = false
@@ -35,7 +36,9 @@ class CheckPythonTask extends BasePipTask {
     void run() {
         PythonExtension ext = project.extensions.findByType(PythonExtension)
         boolean envRequested = ext.scope >= PythonExtension.Scope.VIRTUALENV_OR_USER
-        Virtualenv env = envRequested ? new Virtualenv(project, ext.pythonPath, ext.pythonBinary, ext.envPath) : null
+        Virtualenv env = envRequested
+                // synchronize work dir between python instances
+                ? new Virtualenv(project, ext.pythonPath, ext.pythonBinary, ext.envPath).workDir(getWorkDir()) : null
 
         // use env right ahead (global python could even not exists), but only if allowed by scope
         if (envRequested && env.exists()) {
@@ -68,7 +71,7 @@ class CheckPythonTask extends BasePipTask {
 
     private void checkPython(PythonExtension ext) {
         // important because python could change on second execution
-        Python python = new Python(project, pythonPath, pythonBinary)
+        Python python = new Python(project, pythonPath, pythonBinary).workDir(getWorkDir())
         try {
             python.version
         } catch (ExecException ex) {
@@ -94,7 +97,7 @@ class CheckPythonTask extends BasePipTask {
 
     private void checkPip(PythonExtension ext) {
         // important because python could change on second execution
-        Pip pip = new Pip(project, ext.pythonPath, ext.pythonBinary, false)
+        Pip pip = new Pip(project, ext.pythonPath, ext.pythonBinary, false).workDir(getWorkDir())
         try {
             pip.versionLine
         } catch (PythonExecutionFailed ex) {
@@ -115,7 +118,7 @@ class CheckPythonTask extends BasePipTask {
     }
 
     private boolean checkEnv(Virtualenv env, PythonExtension ext) {
-        Pip pip = new Pip(project, ext.pythonPath, ext.pythonBinary, true)
+        Pip pip = new Pip(project, ext.pythonPath, ext.pythonBinary, true).workDir(getWorkDir())
         if (!pip.isInstalled(env.name)) {
             if (ext.installVirtualenv) {
                 // automatically install virtualenv if allowed (in --user)